requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.3
llama_cpp_python==0.2.53
MarkupSafe==2.1.5
numpy==1.26.3
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-runtime-cu12==12.3.101
packaging==23.2
py-cpuinfo==9.0.0
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sse-starlette==2.0.0
starlette==0.36.3
starlette-context==0.3.6
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.1
